requirements.txt
marimo==0.9.1
python-dotenv
loguru
lancedb==0.14.0
langchain==0.2.14
langchain-community==0.2.12
sentence-transformers[onnx]==3.2.0
groq
polars==1.8.2
pandas==2.2.2
tiktoken
altair